requirements.txt
appdirs==1.4.4
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dateutils==0.6.12
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
frozendict==2.4.7
greenlet==3.3.2
html5==0.0.9
html5lib==1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.8.2
multitasking==0.0.12
newsapi-python==0.2.7
nsepy==0.8
nsetools==2.0.1
numpy==1.26.4
packaging==26.0
pandas==3.0.1
peewee==3.19.0
pillow==12.1.0
platformdirs==4.9.4
protobuf==7.34.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.3.2
scipy==1.17.0
seaborn==0.13.1
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vaderSentiment==3.3.2
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.6
xgboost==2.0.3
yfinance==1.2.0
